*Implementing window buttons *Fix some bugs due to null values being set at different times *Clean up of code by moving the mockOS object and functions to a separate file
