+++ app.js
@@ -7,23 +7,7 @@ let numAppsPerHeight = Math.floor((window.innerHeight - 100) / 125);
 
 
 
-function camelCaseToName(camel){
-    let newName = camel;
-    newName = newName.charAt(0).toUpperCase() + newName.slice(1);
 
-    let slicedName = newName.charAt(0);
-
-    for (let i = 1; i < newName.length; i++){
-        if (newName.charAt(i) === newName.charAt(i).toUpperCase()){
-            slicedName += " ";
-        }
-        slicedName += newName.charAt(i);
-    }
-    return slicedName;
-}
-function clampNum(minimum, base, maximum){
-    return Math.min( Math.max(base, minimum), maximum);
-}
 
 
 
@@ -203,8 +187,7 @@ document.addEventListener("DOMContentLoaded", (event)=>{
      */
     //Clock Functions
 
-    updater = setInterval(mockOS.updateClock, 1000, clock);
-    
+
 
 });
