+++ index.html
@@ -91,6 +91,7 @@
       margin: 2px;
       border-width: 0;
       border-radius: 50%;
+      color: var(--text_color);
     }
     button.TitleBarOptionsButton:hover{
       background-image: linear-gradient(180deg, rgba(255, 255, 255, 0.9), rgba(200,200,200,0.1) 50%);
@@ -126,8 +127,8 @@
       <p class="AppWindowTitleString"></p>
 
       <div class="TitleBarOptions">
-        <button type="button" id="close" class="TitleBarOptionsButton"></button>
-        <button type="button" id="minimize" class="TitleBarOptionsButton"></button>
+        <button type="button" id="close" class="TitleBarOptionsButton">&#x1F5D9;</button>
+        <button type="button" id="minimize" class="TitleBarOptionsButton">&#x2581;</button>
 
       </div>
